+++ src/CircleMarker.js
@@ -3,19 +3,21 @@
 import { circleMarker } from 'leaflet'
 import { PropTypes } from 'react'
 
+import childrenType from './types/children'
 import latlngType from './types/latlng'
 import Path from './Path'
 
 export default class CircleMarker extends Path {
   static propTypes = {
     center: latlngType.isRequired,
+    children: childrenType,
     radius: PropTypes.number,
   };
 
   componentWillMount () {
     super.componentWillMount()
     const { center, ...props } = this.props
-    this.leafletElement = circleMarker(center, props)
+    this.leafletElement = circleMarker(center, this.getOptions(props))
   }
 
   componentDidUpdate (prevProps: Object) {