+++ src/FeatureGroup.js
@@ -3,16 +3,20 @@
 import { featureGroup } from 'leaflet'
 import { PropTypes } from 'react'
 
+import childrenType from './types/children'
 import layerContainerType from './types/layerContainer'
 import Path from './Path'
 
+type FGChild = { layerContainer: Object, popupContainer: Object };
+
 export default class FeatureGroup extends Path {
   static childContextTypes = {
+    children: childrenType,
     layerContainer: layerContainerType,
     popupContainer: PropTypes.object,
   };
 
-  getChildContext () {
+  getChildContext (): FGChild {
     return {
       layerContainer: this.leafletElement,
       popupContainer: this.leafletElement,
@@ -20,7 +24,7 @@ export default class FeatureGroup extends Path {
   }
 
   componentWillMount () {
-    this.leafletElement = featureGroup()
+    this.leafletElement = featureGroup(this.getOptions(this.props))
   }
 
   componentDidMount () {