+++ src/GeoJSON.js
@@ -1,20 +1,22 @@
 /* @flow */
 
-import { geoJson } from 'leaflet'
+import { geoJSON } from 'leaflet'
 import { isFunction } from 'lodash'
 import { PropTypes } from 'react'
 
+import childrenType from './types/children'
 import Path from './Path'
 
-export default class GeoJson extends Path {
+export default class GeoJSON extends Path {
   static propTypes = {
+    children: childrenType,
     data: PropTypes.oneOfType([PropTypes.array, PropTypes.object]).isRequired,
   };
 
   componentWillMount () {
     super.componentWillMount()
     const { data, ...props } = this.props
-    this.leafletElement = geoJson(data, props)
+    this.leafletElement = geoJSON(data, this.getOptions(props))
   }
 
   componentDidUpdate (prevProps: Object) {