+++ src/GridLayer.js
@@ -1,14 +1,22 @@
 /* @flow */
 
+import { gridLayer } from 'leaflet'
 import { PropTypes } from 'react'
 
+import childrenType from './types/children'
 import MapLayer from './MapLayer'
 
-export default class BaseTileLayer extends MapLayer {
+export default class GridLayer extends MapLayer {
   static propTypes = {
+    children: childrenType,
     opacity: PropTypes.number,
     zIndex: PropTypes.number,
   };
+
+  componentWillMount () {
+    super.componentWillMount()
+    this.leafletElement = gridLayer(this.getOptions(this.props))
+  }
 
   componentDidUpdate (prevProps: Object) {
     const { opacity, zIndex } = this.props
@@ -20,7 +28,7 @@ export default class BaseTileLayer extends MapLayer {
     }
   }
 
-  render () {
+  render (): null {
     return null
   }
 }