+++ src/LayerGroup.js
@@ -3,6 +3,7 @@
 import { layerGroup } from 'leaflet'
 
 import layerContainerType from './types/layerContainer'
+
 import MapLayer from './MapLayer'
 
 export default class LayerGroup extends MapLayer {
@@ -10,7 +11,7 @@ export default class LayerGroup extends MapLayer {
     layerContainer: layerContainerType,
   };
 
-  getChildContext () {
+  getChildContext (): { layerContainer: Object } {
     return {
       layerContainer: this.leafletElement,
     }
@@ -18,6 +19,6 @@ export default class LayerGroup extends MapLayer {
 
   componentWillMount () {
     super.componentWillMount()
-    this.leafletElement = layerGroup()
+    this.leafletElement = layerGroup(this.getOptions())
   }
 }