+++ src/Polyline.js
@@ -1,19 +1,25 @@
 /* @flow */
 
 import { polyline } from 'leaflet'
+import { PropTypes } from 'react'
 
+import childrenType from './types/children'
 import latlngListType from './types/latlngList'
 import Path from './Path'
 
 export default class Polyline extends Path {
   static propTypes = {
-    positions: latlngListType.isRequired,
+    children: childrenType,
+    positions: PropTypes.oneOfType([
+      latlngListType,
+      PropTypes.arrayOf(latlngListType),
+    ]).isRequired,
   };
 
   componentWillMount () {
     super.componentWillMount()
     const { positions, ...props } = this.props
-    this.leafletElement = polyline(positions, props)
+    this.leafletElement = polyline(positions, this.getOptions(props))
   }
 
   componentDidUpdate (prevProps: Object) {