+++ src/Rectangle.js
@@ -1,19 +1,23 @@
 /* @flow */
 
 import { rectangle } from 'leaflet'
+import { PropTypes } from 'react'
 
 import boundsType from './types/bounds'
+import childrenType from './types/children'
 import Path from './Path'
 
 export default class Rectangle extends Path {
   static propTypes = {
+    children: childrenType,
     bounds: boundsType.isRequired,
+    popupContainer: PropTypes.object,
   };
 
   componentWillMount () {
     super.componentWillMount()
     const { bounds, ...props } = this.props
-    this.leafletElement = rectangle(bounds, props)
+    this.leafletElement = rectangle(bounds, this.getOptions(props))
   }
 
   componentDidUpdate (prevProps: Object) {