+++ src/ScaleControl.js
@@ -3,6 +3,7 @@
 import { control } from 'leaflet'
 import { PropTypes } from 'react'
 
+import controlPositionType from './types/controlPosition'
 import MapControl from './MapControl'
 
 export default class ScaleControl extends MapControl {
@@ -10,6 +11,7 @@ export default class ScaleControl extends MapControl {
     imperial: PropTypes.bool,
     maxWidth: PropTypes.number,
     metric: PropTypes.bool,
+    position: controlPositionType,
     updateWhenIdle: PropTypes.bool,
   };