+++ src/TileLayer.js
@@ -3,17 +3,21 @@
 import { tileLayer } from 'leaflet'
 import { PropTypes } from 'react'
 
-import BaseTileLayer from './BaseTileLayer'
+import childrenType from './types/children'
+import GridLayer from './GridLayer'
 
-export default class TileLayer extends BaseTileLayer {
+export default class TileLayer extends GridLayer {
   static propTypes = {
+    children: childrenType,
+    opacity: PropTypes.number,
     url: PropTypes.string.isRequired,
+    zIndex: PropTypes.number,
   };
 
   componentWillMount () {
     super.componentWillMount()
     const { url, ...props } = this.props
-    this.leafletElement = tileLayer(url, props)
+    this.leafletElement = tileLayer(url, this.getOptions(props))
   }
 
   componentDidUpdate (prevProps: Object) {