+++ CHANGELOG.md
@@ -1,3 +1,65 @@
+## v1.0.0 (2016-11-14)
+
+Changes from v0.12.3:
+
+- [BREAKING] Updated Leaflet dependency requirement to v1.0.0.
+- [BREAKING] Renamed `GeoJson` to `GeoJSON` to match Leaflet's change.
+- [BREAKING] Removed `MultiPolygon` and `MultiPolyline`, now supported by `Polygon` and `Polyline` respectively.
+- [BREAKING] Removed `CanvasTileLayer`, replaced by `GridLayer`. See [Leaflet's documentation](http://leafletjs.com/reference-1.0.0.html#gridlayer).
+- [BREAKING] Removed `getLeafletElement()` method in `MapComponent` and `MapControl`, deprecated since v0.12.0.
+- [BREAKING] Removed `setIconDefaultImagePath()` helper, set `Leaflet.Icon.Default.imagePath` directly using Leaflet if you need.
+- `Leaflet.Icon.Default.imagePath` is no longer set by default, as per Leaflet v1.0.0 it should be auto-detected.
+- The `Map` component no longer creates an unique ID for its container if none is provided.
+- Added `useFlyTo: bool` property to `Map` to use `flyTo` instead of `setView` and `flyToBounds` instead of `fitBounds` when changing the map's view or bounds ([PR #148](https://github.com/PaulLeCam/react-leaflet/pull/192) by *jgimbel*).
+- Added `GridLayer`.
+- Added `Pane` component introduced in Leaflet v1. ([PR #227](https://github.com/PaulLeCam/react-leaflet/pull/227) by *rjdestigter*).
+- Added `Tooltip` component introduced in Leaflet v1. See [Leaflet's documentation](http://leafletjs.com/reference-1.0.0.html#tooltip) for usage.
+- Added support for IE9 and IE10 ([PR #235](https://github.com/PaulLeCam/react-leaflet/pull/235) by *dettier*).
+- Updated `WMSTileLayer` to support dynamic `url` and parameters.
+- [internal] Fixed Flow errors ([PR #251](https://github.com/PaulLeCam/react-leaflet/pull/251) by *joeyrosztoczy*).
+- [internal] Updated `MapLayer` to only create divs when necessary.
+- [internal] Removed `BaseTileLayer`, replaced by `GridLayer`.
+- [internal] Removed `bower.json` file.
+
+## v1.0.0-rc.3 (2016-11-04)
+
+Updated `WMSTileLayer` to support dynamic `url` property.
+
+## v1.0.0-rc.2 (2016-10-20)
+
+- Added `Pane` component introduced in Leaflet v1. ([PR #227](https://github.com/PaulLeCam/react-leaflet/pull/227) by *rjdestigter*).
+- Added support for IE9 and IE10 ([PR #235](https://github.com/PaulLeCam/react-leaflet/pull/235) by *dettier*).
+- Updated `WMSTileLayer` to support dynamic parameters.
+
+## v1.0.0-rc.1 (2016-09-28)
+
+- [BREAKING] Removed `setIconDefaultImagePath()` helper, set `Leaflet.Icon.Default.imagePath` directly using Leaflet if you need.
+- Updated Leaflet dependency requirement to v1.0.0.
+- `Leaflet.Icon.Default.imagePath` is no longer set by default, as per Leaflet v1.0.0 it should be auto-detected.
+
+## v1.0.0-beta.3 (2016-08-26)
+
+Fixed `Tooltip` removal handling.
+
+## v1.0.0-beta.2 (2016-08-20)
+
+- Added `Tooltip` component introduced in Leaflet v1.0.0-rc.2. See [Leaflet's documentation](http://leafletjs.com/reference-1.0.0.html#tooltip) for usage.
+- Updated Leaflet dependency to v1.0.0-rc.3.
+- Reverted basing components on `React.PureComponent`, they are now based on `React.Component`.
+
+## v1.0.0-beta.1 (2016-07-30)
+
+- [BREAKING] Renamed `GeoJson` to `GeoJSON` to match Leaflet's change.
+- [BREAKING] Removed `MultiPolygon` and `MultiPolyline`, now supported by `Polygon` and `Polyline` respectively.
+- [BREAKING] Removed `CanvasTileLayer`, replaced by `GridLayer`. See [Leaflet's documentation](http://leafletjs.com/reference-1.0.0.html#gridlayer).
+- [BREAKING] Removed `getLeafletElement()` method in `MapComponent` and `MapControl`, deprecated since v0.12.0.
+- Removed `BaseTileLayer`, replaced by `GridLayer`.
+- Removed `bower.json` file.
+- The `Map` component no longer creates an unique ID for its container if none is provided.
+- All components now extend from `React.PureComponent` instead of `React.Component`, as such React v15.3.0+ is now required.
+- Added `GridLayer`.
+- Added `useFlyTo: bool` property to `Map` to use `flyTo` instead of `setView` and `flyToBounds` instead of `fitBounds` when changing the map's view or bounds ([PR #148](https://github.com/PaulLeCam/react-leaflet/pull/192) by *jgimbel*).
+
 ## v0.12.3 (2016-09-21)
 
 Fixed events bindind in UMD build.
@@ -20,7 +82,7 @@ Fixed events bindind in UMD build.
 
 ## v0.11.7 (2016-06-14)
 
-- Fixed removing layers from `LayersControl`.
+Fixed removing layers from `LayersControl`.
 
 ## v0.11.6 (2016-06-09)